graphApi/requirements.txt
aniso8601==7.0.0
anyjson==0.3.3
Django==2.2.6
django-admin-generator==2.2.0
django-filter==2.2.0
django-seed==0.1.9
django-sendsms==0.4
django-tinymce4-lite==1.7.5
django-utils2==2.5.4
Faker==2.0.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
jsmin==2.2.2
promise==2.2.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
